backend done

+++ database/migrations/0001_01_01_000002_create_jobs_table.php
@@ -16,9 +16,9 @@ return new class extends Migration
             $table->string('queue')->index();
             $table->longText('payload');
             $table->unsignedSmallInteger('attempts');
-            $table->unsignedInteger('reserved_at')->nullable();
-            $table->unsignedInteger('available_at');
-            $table->unsignedInteger('created_at');
+            $table->unsignedBigInteger('reserved_at')->nullable();
+            $table->unsignedBigInteger('available_at');
+            $table->unsignedBigInteger('created_at');
         });
 
         Schema::create('job_batches', function (Blueprint $table) {
@@ -30,8 +30,8 @@ return new class extends Migration
             $table->longText('failed_job_ids');
             $table->mediumText('options')->nullable();
             $table->integer('cancelled_at')->nullable();
-            $table->integer('created_at');
-            $table->integer('finished_at')->nullable();
+            $table->unsignedBigInteger('created_at');
+            $table->unsignedBigInteger('finished_at')->nullable();
         });
 
         Schema::create('failed_jobs', function (Blueprint $table) {
